Move check_crate_and_report_errors from check_cmd to nargo::ops::compile
