+++ tooling/nargo/src/ops/mod.rs
@@ -1,7 +1,7 @@
 pub use self::check::check_program;
 pub use self::compile::{
-    collect_errors, compile_contract, compile_program, compile_program_with_debug_instrumenter,
-    compile_workspace, report_errors,
+    check_crate_and_report_errors, collect_errors, compile_contract, compile_program,
+    compile_program_with_debug_instrumenter, compile_workspace, report_errors,
 };
 pub use self::optimize::{optimize_contract, optimize_program};
 pub use self::transform::{transform_contract, transform_program};
